Revert "build(webpack): Change webpack to have smaller chunks and make all entries async (#25566)"
This reverts commit b4fef9f019a907d6eb46c9856f695e1679f35675.
