+++ src/sentry/templates/sentry/bases/react.html
@@ -1,1 +1,2 @@
 {% extends "sentry/base-react.html" %}
+{% load sentry_assets %}
