--- static/app/views/integrationPipeline/init.tsx
@@ -1,5 +1,0 @@
-import 'focus-visible';
-
-import {initializePipelineView} from 'app/bootstrap/initializePipelineView';
-
-initializePipelineView(window.__initialData);
